Split connection diagnostic modules
