+++ apps/mobile/src/lib/connection-diagnostic-types.ts
@@ -1,0 +1,100 @@
+export type ConnectionDiagnosticTrigger =
+	| 'initial-auto-connect'
+	| 'reconnect'
+	| 'manual-diagnostic'
+	| 'command-menu';
+
+export type ConnectionDiagnosticStatus =
+	| 'running'
+	| 'failed'
+	| 'connected'
+	| 'skipped';
+
+export type ConnectionDiagnosticSource =
+	| 'latest-shell'
+	| 'active-connection'
+	| 'saved-entry'
+	| 'tailscale-recovery'
+	| 'reconnect-controller'
+	| 'manual-diagnostic'
+	| 'foreground-service'
+	| 'command-menu';
+
+export type ConnectionDiagnosticConnectionIdentity = {
+	savedConnectionId?: string;
+	connectionId?: string;
+	username?: string;
+	host?: string;
+	port?: number;
+	keyId?: string;
+	useTmux?: boolean;
+	tmuxSessionName?: string;
+};
+
+export type ConnectionDiagnosticError = {
+	name: string;
+	message: string;
+	stack?: string;
+	tag?: string;
+	inner?: unknown;
+};
+
+export type ConnectionDiagnosticEventInput = {
+	type: string;
+	source: ConnectionDiagnosticSource;
+	message?: string;
+	connection?: ConnectionDiagnosticConnectionIdentity;
+	error?: ConnectionDiagnosticError;
+	details?: Record<string, unknown>;
+};
+
+export type ConnectionDiagnosticEvent = ConnectionDiagnosticEventInput & {
+	atMs: number;
+	elapsedMs: number;
+};
+
+export type ConnectionDiagnosticTrace = {
+	id: string;
+	trigger: ConnectionDiagnosticTrigger;
+	reason: string;
+	status: ConnectionDiagnosticStatus;
+	startedAtMs: number;
+	finishedAtMs?: number;
+	events: ConnectionDiagnosticEvent[];
+};
+
+export type ConnectionDiagnosticTraceHandle = {
+	readonly trace: ConnectionDiagnosticTrace;
+	event: (input: ConnectionDiagnosticEventInput) => ConnectionDiagnosticEvent;
+	finish: (status: Exclude<ConnectionDiagnosticStatus, 'running'>) => void;
+};
+
+export type ConnectionDiagnosticAppState = {
+	platformOS: string;
+	pathname?: string;
+	isAutoConnecting: boolean;
+	isReconnecting: boolean;
+	foregroundServiceStarted?: boolean;
+	backgroundWorkAllowed?: boolean;
+	foregroundServiceRequired?: boolean;
+	appActive?: boolean;
+};
+
+export type ConnectionDiagnosticRecorder = {
+	startTrace: (input: {
+		trigger: ConnectionDiagnosticTrigger;
+		reason: string;
+	}) => ConnectionDiagnosticTraceHandle;
+	getLatestTrace: () => ConnectionDiagnosticTrace | null;
+	getHistory: () => ConnectionDiagnosticTrace[];
+	clear: () => void;
+};
+
+export type ConnectionDiagnosticRecorderOptions = {
+	now?: () => number;
+	maxHistory?: number;
+};
+
+export type ConnectionDiagnosticPromptOptions = {
+	appState?: ConnectionDiagnosticAppState;
+};
